Lazy load heavy edit.js blocks + move Inspector control of masonry into own file
