+++ src/blocks/gallery-carousel/edit.js
@@ -16,6 +16,7 @@ import Inspector from './inspector';
 import Swiper from '../../components/swiper';
 import { CarouselGalleryVariationPicker, hasVariationSet } from './variations';
 import { GalleryCarouselContext, GalleryContextProvider } from './context';
+import { parseNavForClass } from './util';
 
 /**
  * WordPress dependencies
@@ -25,20 +26,6 @@ import { Icon } from '@wordpress/icons';
 import { __, sprintf } from '@wordpress/i18n';
 import { ResizableBox, withNotices } from '@wordpress/components';
 import { useCallback, useContext, useEffect, useMemo } from '@wordpress/element';
-
-/**
- * Function will dynamically process a string for use with the `navForClass` block attribute.
- * This attribute should only be truthy when the corresponding `thumbnails` attribute is also set.
- *
- * @constant parseNavForClass
- * @type {Function}
- * @param {boolean} thumbnails This boolean value is equal to props.attributes.thumbnails.
- * @param {string}  clientId   This string value is equal to props.clientId.
- * @return {string} Return parsed class string if thumbnails is truthy or an empty string.
- */
-export const parseNavForClass = ( thumbnails, clientId ) => thumbnails
-	? `has-nav-${ clientId.split( '-' )[ 0 ] }`
-	: '';
 
 const GalleryCarouselEdit = ( props ) => {
 	const { selectBlock } = useDispatch( 'core/block-editor' );
