+++ src/blocks/gallery-carousel/index.js
@@ -4,22 +4,23 @@
 import { GalleryCarouselIcon as icon } from '@godaddy-wordpress/coblocks-icons';
 
 /**
+ * WordPress dependencies
+ */
+import { __ } from '@wordpress/i18n';
+import { Icon, Spinner } from '@wordpress/components';
+import { lazy, Suspense } from '@wordpress/element';
+
+/**
  * Internal dependencies
  */
 import deprecated from './deprecated';
-import edit from './edit';
+const Edit = lazy( () => import( './edit' ) );
+import { GalleryAttributes } from '../../components/block-gallery/shared';
+import { hasFormattingCategory } from '../../utils/block-helpers';
 import metadata from './block.json';
 import save from './save';
 import transforms from './transforms';
 import { variations } from './variations';
-import { GalleryAttributes } from '../../components/block-gallery/shared';
-import { hasFormattingCategory } from '../../utils/block-helpers';
-
-/**
- * WordPress dependencies
- */
-import { __ } from '@wordpress/i18n';
-import { Icon } from '@wordpress/components';
 
 /**
  * Block constants
@@ -55,7 +56,11 @@ const settings = {
 		coBlocksSpacing: true,
 	},
 	transforms,
-	edit,
+	edit: ( props ) => (
+		<Suspense fallback={ <Spinner /> }>
+			<Edit { ...props } />
+		</Suspense>
+	),
 	save,
 	deprecated,
 };
