+++ src/blocks/gallery-collage/index.js
@@ -4,21 +4,22 @@
 import { GalleryCollageIcon as icon } from '@godaddy-wordpress/coblocks-icons';
 
 /**
- * Internal dependencies.
- */
-import deprecated from './deprecated';
-import edit from './edit';
-import metadata from './block.json';
-import save from './save';
-import transforms from './transforms';
-import { GalleryAttributes } from '../../components/block-gallery/shared';
-import { hasFormattingCategory } from '../../utils/block-helpers';
-
-/**
  * WordPress dependencies.
  */
 import { __ } from '@wordpress/i18n';
-import { Icon } from '@wordpress/components';
+import { Icon, Spinner } from '@wordpress/components';
+import { lazy, Suspense } from '@wordpress/element';
+
+/**
+ * Internal dependencies.
+ */
+import deprecated from './deprecated';
+const Edit = lazy( () => import( './edit' ) );
+import { GalleryAttributes } from '../../components/block-gallery/shared';
+import { hasFormattingCategory } from '../../utils/block-helpers';
+import metadata from './block.json';
+import save from './save';
+import transforms from './transforms';
 
 /**
  * Block constants.
@@ -85,7 +86,11 @@ const settings = {
 	},
 	attributes,
 	transforms,
-	edit,
+	edit: ( props ) => (
+		<Suspense fallback={ <Spinner /> }>
+			<Edit { ...props } />
+		</Suspense>
+	),
 	save,
 	deprecated,
 };
