+++ src/blocks/gallery-masonry/edit.js
@@ -2,21 +2,18 @@
  * External dependencies
  */
 import classnames from 'classnames';
-import GutterControl from '../../components/gutter-control/gutter-control';
 import { GalleryMasonryIcon as icon } from '@godaddy-wordpress/coblocks-icons';
 import { concat, find } from 'lodash';
 
 /**
  * WordPress dependencies
  */
+import { __ } from '@wordpress/i18n';
 import { compose } from '@wordpress/compose';
 import { createBlobURL } from '@wordpress/blob';
 import { createBlock } from '@wordpress/blocks';
-import { store as noticesStore } from '@wordpress/notices';
-import { View } from '@wordpress/primitives';
-import { __, _x, sprintf } from '@wordpress/i18n';
-import { BaseControl, Icon, PanelBody, RangeControl, SelectControl, Spinner, ToggleControl, withNotices } from '@wordpress/components';
-import { store as blockEditorStore, InspectorControls, MediaPlaceholder, useBlockProps } from '@wordpress/block-editor';
+import { store as blockEditorStore, MediaPlaceholder, useBlockProps } from '@wordpress/block-editor';
+import { Icon, withNotices } from '@wordpress/components';
 import { Platform, useEffect, useMemo } from '@wordpress/element';
 import { useDispatch, useSelect } from '@wordpress/data';
 
@@ -26,21 +23,13 @@ import { useDispatch, useSelect } from '@wordpress/data';
 import Controls from './controls.js';
 import Gallery from '../../components/block-gallery/gallery';
 import GutterWrapper from '../../components/gutter-control/gutter-wrapper';
+import Inspector from './inspector';
+import { LINK_DESTINATION_NONE } from '../../components/block-gallery/constants';
 import { pickRelevantMediaFiles } from '../../components/block-gallery/shared-helpers';
 import useGetMedia from '../../components/block-gallery/use-get-media';
 import useGetNewImages from '../../components/block-gallery/use-get-new-images';
-import useImageSizes from '../../components/block-gallery/use-image-sizes';
-import { getHrefAndDestination, getImageSizeAttributes, getUpdatedLinkTargetSettings } from '../../components/block-gallery/utils';
-import { LINK_DESTINATION_ATTACHMENT, LINK_DESTINATION_MEDIA, LINK_DESTINATION_NONE } from '../../components/block-gallery/constants';
+import { getHrefAndDestination, getUpdatedLinkTargetSettings } from '../../components/block-gallery/utils';
 
-const linkOptions = [
-	{ label: __( 'Attachment Page', 'coblocks' ), value: LINK_DESTINATION_ATTACHMENT },
-	{ label: __( 'Media File', 'coblocks' ), value: LINK_DESTINATION_MEDIA },
-	{
-		label: _x( 'None', 'Media item link option' ),
-		value: LINK_DESTINATION_NONE,
-	},
-];
 const ALLOWED_MEDIA_TYPES = [ 'image' ];
 
 const PLACEHOLDER_TEXT = Platform.isNative
@@ -61,13 +50,9 @@ function GalleryEdit( props ) {
 
 	const {
 		filter,
-		imageCrop,
 		linkTarget,
 		linkTo,
 		sizeSlug,
-		lightbox,
-		gutter,
-		gutterCustom,
 		radius,
 	} = attributes;
 
@@ -77,15 +62,13 @@ function GalleryEdit( props ) {
 		replaceInnerBlocks,
 		updateBlockAttributes,
 	} = useDispatch( blockEditorStore );
-	const { createSuccessNotice } = useDispatch( noticesStore );
 
-	const { getBlock, getSettings, preferredStyle } = useSelect( ( select ) => {
+	const { getBlock, preferredStyle } = useSelect( ( select ) => {
 		const settings = select( blockEditorStore ).getSettings();
 		const preferredStyleVariations =
 			settings.__experimentalPreferredStyleVariations;
 		return {
 			getBlock: select( blockEditorStore ).getBlock,
-			getSettings: select( blockEditorStore )?.getSettings || {},
 			preferredStyle: preferredStyleVariations?.value?.[ 'core/image' ],
 		};
 	}, [] );
@@ -182,12 +165,6 @@ function GalleryEdit( props ) {
 			} );
 		}
 	}, [ linkTo ] );
-
-	const imageSizeOptions = useImageSizes(
-		imageData,
-		isSelected,
-		getSettings
-	);
 
 	/**
 	 * Determines the image attributes that should be applied to an image block
@@ -332,102 +309,6 @@ function GalleryEdit( props ) {
 		noticeOperations.createErrorNotice( message );
 	}
 
-	function setLinkTo( value ) {
-		setAttributes( { linkTo: value } );
-		const changedAttributes = {};
-		const blocks = [];
-		getBlock( clientId ).innerBlocks.forEach( ( block ) => {
-			blocks.push( block.clientId );
-			const image = block.attributes.id
-				? find( imageData, { id: block.attributes.id } )
-				: null;
-			changedAttributes[ block.clientId ] = getHrefAndDestination(
-				image,
-				value
-			);
-		} );
-		updateBlockAttributes( blocks, changedAttributes, true );
-		const linkToText = [ ...linkOptions ].find(
-			( linkType ) => linkType.value === value
-		);
-
-		createSuccessNotice(
-			sprintf(
-				/* translators: %s: image size settings */
-				__( 'All gallery image links updated to: %s', 'coblocks' ),
-				linkToText.label
-			),
-			{
-				id: 'gallery-attributes-linkTo',
-				type: 'snackbar',
-			}
-		);
-	}
-
-	function toggleImageCrop() {
-		setAttributes( { imageCrop: ! imageCrop } );
-	}
-
-	function getImageCropHelp( checked ) {
-		return checked
-			? __( 'Thumbnails are cropped to align.', 'coblocks' )
-			: __( 'Thumbnails are not cropped.', 'coblocks' );
-	}
-
-	function toggleOpenInNewTab( openInNewTab ) {
-		const newLinkTarget = openInNewTab ? '_blank' : undefined;
-		setAttributes( { linkTarget: newLinkTarget } );
-		const changedAttributes = {};
-		const blocks = [];
-		getBlock( clientId ).innerBlocks.forEach( ( block ) => {
-			blocks.push( block.clientId );
-			changedAttributes[ block.clientId ] = getUpdatedLinkTargetSettings(
-				newLinkTarget,
-				block.attributes
-			);
-		} );
-		updateBlockAttributes( blocks, changedAttributes, true );
-		const noticeText = openInNewTab
-			? __( 'All gallery images updated to open in new tab', 'coblocks' )
-			: __( 'All gallery images updated to not open in new tab', 'coblocks' );
-		createSuccessNotice( noticeText, {
-			id: 'gallery-attributes-openInNewTab',
-			type: 'snackbar',
-		} );
-	}
-
-	function updateImagesSize( newSizeSlug ) {
-		setAttributes( { sizeSlug: newSizeSlug } );
-		const changedAttributes = {};
-		const blocks = [];
-		getBlock( clientId ).innerBlocks.forEach( ( block ) => {
-			blocks.push( block.clientId );
-			const image = block.attributes.id
-				? find( imageData, { id: block.attributes.id } )
-				: null;
-			changedAttributes[ block.clientId ] = getImageSizeAttributes(
-				image,
-				newSizeSlug
-			);
-		} );
-		updateBlockAttributes( blocks, changedAttributes, true );
-		const imageSize = imageSizeOptions.find(
-			( size ) => size.value === newSizeSlug
-		);
-
-		createSuccessNotice(
-			sprintf(
-				/* translators: %s: image size settings */
-				__( 'All gallery image sizes updated to: %s', 'coblocks' ),
-				imageSize.label
-			),
-			{
-				id: 'gallery-attributes-sizeSlug',
-				type: 'snackbar',
-			}
-		);
-	}
-
 	const hasImages = !! images.length;
 	const hasImageIds = hasImages && images.some( ( image ) => !! image.id );
 	const imagesUploading = images.some(
@@ -457,16 +338,6 @@ function GalleryEdit( props ) {
 		/>
 	);
 
-	const getLightboxHelp = ( checked ) => {
-		return checked
-			? __( 'Image lightbox is enabled.', 'coblocks' )
-			: __( 'Toggle to enable the image lightbox.', 'coblocks' );
-	};
-
-	const setRadiusTo = ( value ) => {
-		setAttributes( { radius: value } );
-	};
-
 	const blockProps = useBlockProps( {
 		className: classnames( className, {
 			[ `has-filter-${ filter }` ]: filter !== 'none',
@@ -474,77 +345,14 @@ function GalleryEdit( props ) {
 		} ),
 	} );
 
-	const hasLinkTo = linkTo && linkTo !== 'none';
-
 	return (
 		<>
-			<Controls { ...props } innerBlocks={ images } />
-			<InspectorControls>
-				<PanelBody title={ __( 'Gallery settings', 'coblocks' ) }>
-					<ToggleControl
-						checked={ !! imageCrop }
-						help={ getImageCropHelp }
-						label={ __( 'Crop images', 'coblocks' ) }
-						onChange={ toggleImageCrop }
-					/>
-					<SelectControl
-						hideCancelButton={ true }
-						label={ __( 'Link to', 'coblocks' ) }
-						onChange={ setLinkTo }
-						options={ linkOptions }
-						value={ linkTo }
-					/>
-					{ hasLinkTo && (
-						<ToggleControl
-							checked={ linkTarget === '_blank' }
-							label={ __( 'Open in new tab', 'coblocks' ) }
-							onChange={ toggleOpenInNewTab }
-						/>
-					) }
-					{ imageSizeOptions?.length > 0 && (
-						<SelectControl
-							hideCancelButton={ true }
-							label={ __( 'Image size', 'coblocks' ) }
-							onChange={ updateImagesSize }
-							options={ imageSizeOptions }
-							value={ sizeSlug }
-						/>
-					) }
-					{ Platform.isWeb && ! imageSizeOptions && (
-						<BaseControl className={ 'gallery-image-sizes' }>
-							<BaseControl.VisualLabel>
-								{ __( 'Image size', 'coblocks' ) }
-							</BaseControl.VisualLabel>
-							<View className={ 'gallery-image-sizes__loading' }>
-								<Spinner />
-								{ __( 'Loading options…', 'coblocks' ) }
-							</View>
-						</BaseControl>
-					) }
-
-					<GutterControl { ...props } />
-
-					{ gutter !== 'custom' && gutterCustom !== 0 &&
-					<RangeControl
-						aria-label={ __( 'Add rounded corners to the gallery items.', 'coblocks' ) }
-						label={ __( 'Rounded corners', 'coblocks' ) }
-						max={ 20 }
-						min={ 0 }
-						onChange={ setRadiusTo }
-						step={ 1 }
-						value={ radius }
-					/>
-					}
-
-					<ToggleControl
-						checked={ !! lightbox }
-						help={ getLightboxHelp }
-						label={ __( 'Lightbox', 'coblocks' ) }
-						onChange={ () => setAttributes( { lightbox: ! lightbox } ) }
-					/>
-
-				</PanelBody>
-			</InspectorControls>
+			{ isSelected && (
+				<>
+					<Controls { ...props } innerBlocks={ images } />
+					<Inspector { ...props } />
+				</>
+			) }
 			{ noticeUI }
 			<GutterWrapper { ...attributes }>
 				<Gallery
