+++ src/blocks/gallery-masonry/index.js
@@ -4,21 +4,22 @@
 import { GalleryMasonryIcon as icon } from '@godaddy-wordpress/coblocks-icons';
 
 /**
+ * WordPress dependencies
+ */
+import { __ } from '@wordpress/i18n';
+import { Icon, Spinner } from '@wordpress/components';
+import { lazy, Suspense } from '@wordpress/element';
+
+/**
  * Internal dependencies
  */
 import deprecated from './deprecated';
-import edit from './edit-wrapper';
+const Edit = lazy( () => import( './edit' ) );
 import { GalleryAttributes } from '../../components/block-gallery/shared';
 import { hasFormattingCategory } from '../../utils/block-helpers';
 import metadata from './block.json';
 import save from './save';
 import transforms from './transforms';
-
-/**
- * WordPress dependencies
- */
-import { __ } from '@wordpress/i18n';
-import { Icon } from '@wordpress/components';
 
 /**
  * Block constants
@@ -36,7 +37,11 @@ const settings = {
 	deprecated,
 	/* translators: block description */
 	description: __( 'Display multiple images in an organized masonry gallery.', 'coblocks' ),
-	edit,
+	edit: ( props ) => (
+		<Suspense fallback={ <Spinner /> }>
+			<Edit { ...props } />
+		</Suspense>
+	),
 	example: {
 		innerBlocks: [
 			{
